refactor

+++ src/super_segment/hash_index/mod.rs
@@ -8,7 +8,7 @@ const MARKER_CONFLICT: u8 = u8::MAX;
 
 // NOTE: We know the hash index has a bucket count <= u8
 #[allow(clippy::cast_possible_truncation)]
-fn calculate_bucket_position(key: &[u8], bucket_count: u8) -> usize {
+fn calculate_bucket_position(key: &[u8], bucket_count: u32) -> usize {
     let hash = xxh3_64(key);
     (hash % u64::from(bucket_count)) as usize
 }
